Move user password to library
